requirements.txt
openai>=1.0.0
requests
python-dotenv
pyyaml>=1.0.0
